inference/requirements.txt
kdtree
numpy
Pillow
scipy
tqdm
